+++ CHANGES.md
@@ -13,6 +13,33 @@ development source code and as such may not be routinely kept up to date.
 
 # __NEXT__
 
+## Improvements
+
+* Use of an alternate Conda package repository is now supported during
+  `nextstrain setup conda` and `nextstrain update conda` if you cannot or do
+  not want to use the default package repository hosted by Anaconda.  Set the
+  [`NEXTSTRAIN_CONDA_CHANNEL_ALIAS`][] environment variable to the base URL of
+  the repository.  This corresponds to the [`channel_alias` Conda config
+  setting][].
+  ([#436](https://github.com/nextstrain/cli/pull/436))
+
+* The Conda runtime no longer requires Rosetta 2 for macOS running on aarch64
+  (aka arm64, Apple Silicon, M1/M2/…) hardware.  This improves performance when
+  using the runtime.  Run `nextstrain update conda` to receive the update.
+  ([#436](https://github.com/nextstrain/cli/pull/436))
+
+[`NEXTSTRAIN_CONDA_CHANNEL_ALIAS`]: https://docs.nextstrain.org/projects/cli/en/__NEXT__/runtimes/conda/#envvar-NEXTSTRAIN_CONDA_CHANNEL_ALIAS
+[`channel_alias` Conda config setting]: https://docs.conda.io/projects/conda/en/latest/user-guide/configuration/settings.html#set-ch-alias
+
+## Development
+
+* The `NEXTSTRAIN_CONDA_MICROMAMBA_VERSION` environment variable is no longer
+  supported (i.e. for use with `nextstrain setup conda`).  Please use
+  [`NEXTSTRAIN_CONDA_MICROMAMBA_URL`][] instead.
+  ([#436](https://github.com/nextstrain/cli/pull/436))
+
+[`NEXTSTRAIN_CONDA_MICROMAMBA_URL`]: https://docs.nextstrain.org/projects/cli/en/__NEXT__/runtimes/conda/#envvar-NEXTSTRAIN_CONDA_MICROMAMBA_URL
+
 
 # 10.0.0 (7 May 2025)